Basic template file made for layout
Renamed helpers

Add a helper to print out data in the template

Print out all building names

Created a minimum detail template

Implement a MVC pattern. Moved the homepage function to buildingcontroller.js

Create a buildingcard mixin

Create and render plain detailpage route

Added setup info in the readme
