+++ README.md
@@ -1,10 +1,27 @@
 # performance-matters-server-side
 
+Try to use only ES5 (will use template literals somethimes cus..... dang concatenation)
+
 ## Table of Contents
 
+### Getting started
+
+If you want to work on this project, follow these steps:
+1. First we clone (fork if you want) the repo.
+	Run `git clone https://github.com/kyunwang/performance-matters-server-side.git` in your terminal
+2. `cd` to the repo and run `npm install` to install the dependencies
+3. Run `npm start` to run the server
+4. Go to `http://localhost:3500/`
+5. Have fun!
+
+## Setup server
+Ofcourse I need to setup the server first. I did not remember how to do that on top of my hat(head) - You got it? 😏
+
+I used [this][express-base] simple server as a base and also used the `request` package example.
 
 
-## Start
+
+
 Overview for now
 
 Thinking about what to do....
@@ -17,7 +34,6 @@ Creating a setup with packages I think I might need
 - eslint
 - body-parser
 
-Took the express base from this link [here][express-base]
 
 ## Core functionality
 Thinking about wha to do and how..... want to use wafs man....
@@ -37,6 +53,9 @@ And The build and demolish years, I hard coded in it.
 Add older script from client side version a.k.a. the helpers and made the query to the sparql/wikidata endpoint
 
 Todo: clean helpers and check what is needed. Know more....
+
+
+## Todo
