+++ views/details.pug
@@ -1,0 +1,5 @@
+extends layout
+
+block content
+	h1 The detailpage
+	h3= building.itemLabel.value
