+++ views/layout.pug
@@ -1,0 +1,14 @@
+
+html
+	head
+		title Religious Buildings Amsterdam
+		link(rel='stylesheet', href='/public/style.css')
+		meta(name="viewport" content="width=device-width, initial-scale=1")
+	body
+		block header 
+			header This is the header
+		main 
+			block content
+				p Here should be content
+		block footer
+			footer This is the footer
